🚀 | Merged ServerInfo and UserInfo -> Info
